Renamed after_initialize to afterInitialize.
Following
http://framework.zend.com/manual/1.11/en/coding-standard.naming-conventions.html

"Function names may only contain alphanumeric characters. Underscores are not
permitted."

"Function names must always start with a lowercase letter. When a function name
consists of more than one word, the first letter of each new word must be
capitalized. This is commonly called 'camelCase' formatting."

"For methods on objects that are declared with the "private" or "protected"
modifier, the first character of the method name must be an underscore. This is
the only acceptable application of an underscore in a method name."